(refactor) Update DB models to single file
