+++ main.go
@@ -26,7 +26,7 @@ func main() {
    *------------------------------------*/
   v := api.PathPrefix("/videos").Subrouter()
   v.Methods("GET").Path("/{url}").HandlerFunc(routes.GetVideo)
-  v.Methods("POST").Path("/convert").HandlerFunc(routes.ConvertVideo)
+  v.Methods("POST").Path("/convert").HandlerFunc(routes.ProcessVideo)
   v.Methods("POST").HandlerFunc(routes.CreateVideo)
 
   /*-------------------------------------
