+++ server/src/db/models.go
@@ -1,4 +1,4 @@
-package videos
+package db
 
 import (
   "time"
